Fix regex in 'Customizing Topic Auto-Creation' docs
This topic creation configuration property is a regex and the example in the docs doesn't match the description of what the example is meant to do.

The description says it matches "all topics that start with `dbserver1.inventory.`" but it does not. It matches an arbitrary number of `.` characters at the end. This change fixes the regex to match what the description says.